Added a CreateObject class, some new model tests and tidied up
The CreateObject class allows you to click the left mouse button and it will create the currently selected object at the camera position. 0,1,2 on the numpad for plane, cube and aeroplane respectively. + and - also scales the objects that you create. Added some models with textures and tidied up by removing old code